Add missing bracket
