+++ src/multimodelplaystockTask1.java
